move lookup trait into interpreter, and add two more functions for readability
